[optimize]build userController
